@@ -17,23 +17,23 @@ environment:
   sdk: ">=2.1.0 <3.0.0"
 
 dependencies:
-  curved_navigation_bar: ^0.3.2
+  curved_navigation_bar: ^1.0.0
   flutter:
     sdk: flutter
-  intl: ^0.16.1
-  firebase_auth: ^0.14.0+5
-  cloud_firestore: ^0.13.5
-  firebase_core: ^0.4.2+1
-  provider: ^3.1.0
+  intl: ^0.17.0
+  firebase_auth: ^1.4.1
+  cloud_firestore: ^2.2.2
+  firebase_core: ^1.3.0
+  provider: ^5.0.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  flutter_spinkit: "^4.1.2"
-  cupertino_icons: ^0.1.2
-  shared_preferences: ^0.5.7+2
+  flutter_spinkit: ^5.0.0
+  cupertino_icons: ^1.0.3
+  shared_preferences: ^2.0.6
   progress_dialog: ^1.2.2
-  font_awesome_flutter: ^8.8.1
-  firebase_messaging: ^6.0.16
-  url_launcher: ^5.4.10
+  font_awesome_flutter: ^9.1.0
+  firebase_messaging: ^10.0.2
+  url_launcher: ^6.0.6
 
 dev_dependencies:
   flutter_test:
